[build-troubleshooting] Move requires of 'rake/clean' to files that need them
I think this is what has been failing the builds.

I can see that this would have been an issue previously, but I am not
sure why this only came to light after we started using Cucumber.

I was never able to reproduce these failures locally.

About this change: I am trying to be move the 'dependency' of the
require to the file that needs it, rather than having the dependency
be required in by someone (reveal-ck/bin) who shouldn't be in the
business of understand dependencies of things like rakelib/reveal.rake
or rakelib/slides.html.rake.