[cleaning-up-rakefiles] Slim down all rake files into one.
I can't see it in this project, directly, but the fact that my
rakefiles were all getting loaded and I don't distinguish between
rakefiles needed for 'development' (of reveal-ck) as opposed to
'runtime' (like running reveal-ck generate) means that I was blowing
up trying to run `reveal-ck generate` in a child project because that
child project doesn't have access to a development dependency, such as
cucumber.

Anyway-- for now-- this commit reduces the number of files that are
being loaded in reveal-ck/bin while going green-to-green.