[removing-jeweler] Bumped to 0.1.2
Whether using Jeweler or not, I previously had a runtime dependency on
haml, but this really needed to be haml ~> 3.1.x. As it stands right
now, if you grab reveal-ck 0.1.1 from rubygems and do a fresh bundle,
you'll grab haml 4.x and things don't work.

So-- there are no new features or traditional bug fixes in 0.1.2, but
it should be back to normal again.