[renames] Builders go to builders/
Some parts of RevealCK are associated with creating content within
files, or directories, on the file system. These are known as
"Builders," because they are concerned with translating instructions and
things in memory onto the file system.

That said, any RevealCK file associated with building content now lives
in he module "RevealCK::Builders" and consequently in the directory
"lib/reveal-ck/builders."

In addition to this move, some files were renamed such that their name
describes what they will ultimately build:

- `SlideBuilder` is now `IndexHtml`. It knows how to build the core
  reveal.js index.html.

- `SlidesHtmlBuilder` is now `SlidesHtml`. It knows how to build the
  intermediate slides.html (just your presentation content, no reveal.js
  content).

- `PresentationBuilder` is now `PackagedSlides`. It knows how to bundle
  up the content that goes along with a reveal.js presentation into a
  single directory.

There should be no functional changes in this commit.