[config-and-dsl] PackagedSlides Changes
PackageSlides no longer works with an explicit Config.

It used to work with a file (which it would turn into a Presentation)
and a Config. Previously, both the Config object and the Presentation
contained metadata about the slides. The drive behind this refactoring
is that this metadata will be merged together before we reach
PackagedSlides and PackageSlides will be less involved with this idea.

- Presentation's #add_slide renamed to #add

- Presentation's #content renamed to #html

- Presentation now has #from_template, #from_dsl, and
  load. This logic used to live in builders/SlidesHtml.

- Fixed issue within PresentationDSL where unset attributes were nil,
  and overwriting previously set config.