[tools] Started using rubocop
The current plan of action is to generally accept all defaults and see
what new stuff I learn about ruby.

At present, the two current variations from the default are:

* The default cyclomatic complexity is too high, so go lower.
* Don't want to document my specs. I often create modules just for
  namespacing a describe.

There are hundreds of problems as it stands. Next couple commits should
bring that down and/or introduce new variations.