[index.html.erb] Rewrote IndexHtml from an erb approach
Previously, it was unpacking whatever reveal.js came with, and
transforming that into the presentation of the user.

This came from the mindset where I thought I'd be upgraded more
frequently, and I wanted to "transform whatever reveal.js' index" looked
like into the final result.

No more-- with this approach we'll explicitly maintain an index.html.erb
that may need to be updated each time the embedded reveal.js is
embedded. But-- the thinking is that this is just fine because it
doesn't happen very often and reveal.js isn't changing very much.