[config-refactor] Config is now a Class
Previously, Config was a module that mixed into a few classes to provide
general configuration stuff.

However-- I found this was overly complicated-- in short, a Config
should represent a Hash that gets values as follows:

1. Start with sensible defaults
2. Update entries with values from config.yml (if any)
3. Update entries as slide creation progresses

These entries should be available for reading (or writing) at any point
during the creation process.

And so-- in this first pass-- a Config is a concrete object. There's a
single instance that's passed around, and this instance has settings and
getters that are reasonable for a first pass.