[cleanup] Refactored spec for CopyFilesTask
I think this is better because it presumes *less* about the ultimate
destination of the messages and is instead based on only verifying the
"outgoing" message initiation on CopyFilesTask.

It does this because it assumes that a CopyFilesTask is "RakeAware" and
"RakeAware" has independently been verified to work with a
Rake::Application.

It also reads better.