[cleanup] Correct dependency setup in CreateIndexHtml
Creating the index.html file should have a dependency on the
slides_html, not the #{output_dir}/slides.html. The fact that reveal-ck
wants #{output_dir}/slides.html to exist does not need to be coupled to
creating #{output_dir}/index.html.

When I took a closer look, I saw a discrepancy with the usage of
slides_file and slides_html. I corrected this, and here's my intention:

slides_file: This could be markdown, haml, whatever-- it's not yet
             HTML. It must be transformed into HTML before it can be
             used.

slides_html: This is HTML. It might've started as HTML, or ERB, or HAML,
             or whatever-- but it's now firmly HTML and it is the thing
             that is in-lined into index.html.

Note to future self: I only realized any of this because I wrote a spec
that had different (and more pure) requirements about building an
index.html.