[upcoming] Add <code class="XXX"> when using triple fenced blocks
I recently provided block_code which was a ruby implementation of code
that was formerly in C.

And so.. I think this change gets me back to its "normal" behavior.

However-- reveal.js uses highlight.js which ignores whatever the class
is on the <code>.. so.. I'm not sure how helpful this is.