[vertical-markdown] Changed up pre/post process algorithm
Previously, we'd let the markdown processor take "---" and create an
"<hr/>" out of it.. and then we'd go to work with what was left.

Now, we explicitly seek out "---" and switch it to something the
markdown processor won't touch and doesn't care about-- in the
preprocessing stage: a boring old <div/>.

This is nicer.. but it's happening because I'm planning to use "***" as
the separator for vertical slides, and both --- and *** will end up as
"<hr/>"s.

If that happens-- I won't know which is which in the postproces step.