[dependencies] Futzing with Guard
reveal-ck used to be on a version where Guard was naturally
asynchronous, and this is no longer the case.

Starting guard in a distinct thread solves this (I think).

However-- guard no longer functions the way it used to with
"no_interactions: true", and it doesn't do anything helpful when
present.

So-- for now-- no_interactions isn't specified and it defaults to
false. But, we now have the guard/pry console in the middle of reveal-ck
when it's not needed:

```
------------------------------------------------------------
[ reveal-ck ] Serving up slide content in 'slides/'.
[ reveal-ck ] Open your browser to 'http://localhost:10000'.
[ reveal-ck ] Press CTRL-C to stop.
------------------------------------------------------------
[ reveal-ck ] Getting Ready to Reload Browsers.
[ reveal-ck ] Getting Ready to Rebuild Slides.
[   reload  ] Using Guardfile at /projects/reveal-ck/files/reveal-ck/Guardfile.
[ reveal-ck ] Starting Webserver.
[   reload  ] LiveReload is waiting for a browser to connect.
[   reload  ] Guard is now watching at '/projects/reveal-ck/tmp/aruba'
[1] guard(main)>
[1] guard(main)>
```