[gem-upgrade] Upgrade to latest versions of all gems
This was accomplished by modifying the gemspec and allowing all ranges
to be open, then running a `bundle install`.

With this new set of gems, I then get rubocop, specs, and features to
pass.

Finally, I modify the gemspec so that no ranges are open-- every runtime
dependency is an explicit version and the development dependencies are
increased to be pessismistic at the "minor."