[fix-build] Remove flakey spec
This is a legit spec.. but it's always been a bit flakey and seems to
have gotten worse.

Now that I've been away from ruby I don't know how to easily fix it,
and, frankly, it seems a little far fetched to be testing this in a
"unit test."

This change will cause coverage to dip below 100%.. but what are you
gonna do?