Explicitly update a bunch of libraries
This includes rack and haml-- both of which have security
issues open against them.

In the past, I've seen reveal-ck closer to "an application" than a
library, and so, I've locked versions of libraries in rather than
using ">=" or "~>."

I might not change this now-- when I allowed library versions to
fluctuate a library could produce a change that wouldn't provide
any real benefit to reveal-ck itself but it would cause a problem.

I might change this now-- I'm not updating reveal-ck that much
anymore and if something is discovered that's problematic people
can "get onto the next" w/o having to depend on me.